python/fastapi/mongodb/pyproject.toml
[tool.poetry]
name = "mongodb"
version = "0.1.0"
description = ""
authors = ["Seungyeop Lee <xbeast7@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = { extras = ["standard"], version = "^0.115.0" }
pydantic-settings = "^2.5.2"
motor = "^3.6.0"
beanie = "^1.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"